django-back/requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
filelock==3.4.2
Markdown==3.3.6
platformdirs==2.4.1
psycopg2==2.9.3
pyserial==3.5
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.0
